Sat, Feb 4, 2023, 6:55 PM +08:00

+++ filter-more/helpers.c
@@ -143,9 +143,9 @@ void edges(int height, int width, RGBTRIPLE image[height][width])
                 }
             }
 
-            image[i][j].rgbtRed = round(sqrt(pow((float) Gx.rgbtRed, 2) + pow(Gy.rgbtRed, 2)));
-            image[i][j].rgbtGreen = round(sqrt(pow((float) Gx.rgbtGreen, 2) + pow(Gy.rgbtGreen, 2)));
-            image[i][j].rgbtBlue = round(sqrt(pow((float) Gx.rgbtBlue, 2) + pow(Gy.rgbtBlue, 2)));
+            image[i][j].rgbtRed = round(sqrt(pow((double) Gx.rgbtRed, 2) + pow(Gy.rgbtRed, 2)));
+            image[i][j].rgbtGreen = round(sqrt(pow((double) Gx.rgbtGreen, 2) + pow(Gy.rgbtGreen, 2)));
+            image[i][j].rgbtBlue = round(sqrt(pow((double) Gx.rgbtBlue, 2) + pow(Gy.rgbtBlue, 2)));
 
             if (image[i][j].rgbtRed > 255)
             {
